make HttpContentProcessor an interface again
